Revert "test: avoid circular dependency with core/test crates"
This reverts commit 9bc3b1b66eebf18cdad351f545ba0598fd296d78.
