+++ crates/test/Cargo.toml
@@ -5,10 +5,6 @@ edition = "2021"
 publish = false
 
 [dependencies]
-deltalake-core = { version = "0.26.0", path = "../core", features = [
-    "integration_test",
-] }
-
 arrow-array = { workspace = true, features = ["chrono-tz"] }
 arrow-cast = { workspace = true }
 arrow-ord = { workspace = true }
@@ -19,18 +15,19 @@ parquet = { workspace = true, features = ["async", "object_store"] }
 bytes = { workspace = true }
 chrono = { workspace = true, default-features = false, features = ["clock"] }
 delta_kernel = { workspace = true }
-object_store = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-thiserror = { workspace = true }
-url = { workspace = true }
-
+deltalake-core = { version = "0.26.0", path = "../core" }
 dotenvy = "0"
 fs_extra = "1.3.0"
 futures = { version = "0.3" }
+object_store = { workspace = true }
 rand = "0.8"
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
 tempfile = "3"
+thiserror = { workspace = true }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
+url = { workspace = true }
 
 [features]
 default = []
+datafusion = ["deltalake-core/datafusion"]
